requirements.txt
pillow
numpy
trimesh
matplotlib
scipy